AEGIR_PROOF_OF_CONCEPT_v1.0_FINAL/requirements.txt
# Aegir Smart Reconnaissance Platform
# Core dependencies
aiohttp>=3.8.0,<4.0.0
httpx>=0.24.0,<1.0.0
asyncio>=3.4.3
dnspython>=2.3.0,<3.0.0

# Web scraping and parsing
beautifulsoup4>=4.11.0,<5.0.0
lxml>=4.9.0,<5.0.0

# Screenshot capture
playwright>=1.35.0,<2.0.0

# Data handling
dataclasses>=0.6; python_version<"3.7"
typing-extensions>=4.5.0,<5.0.0

# JSON and configuration
pyyaml>=6.0,<7.0.0

# Development and testing
pytest>=7.3.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
black>=23.0.0,<24.0.0
flake8>=6.0.0,<7.0.0

# Optional: Enhanced features
rich>=13.0.0,<14.0.0  # Better console output
click>=8.1.0,<9.0.0    # CLI improvements 